Refactor welcome wizard naming to more common wizard
[MAILPOET-1732]
