+++ assets/js/src/wizard/create_sender_settings.jsx
