+++ assets/js/src/wizard/steps/migrated_user_step.jsx
