+++ assets/js/src/wizard/steps/sender_step.jsx
