+++ assets/js/src/wizard/steps/step_layout.jsx
