+++ assets/js/src/wizard/steps/usage_tracking_step.jsx
