+++ assets/js/src/wizard/welcome_wizard_controller.jsx
