+++ views/welcome_wizard.html
@@ -15,7 +15,7 @@
   var admin_email = <%= json_encode(admin_email) %>;
 </script>
 
-<div id="welcome_wizard_container"></div>
+<div id="mailpoet_wizard_container"></div>
 
 <div class="welcome_wizard_video">
     <iframe width="1" height="1" src="https://player.vimeo.com/video/279123953" frameborder="0"></iframe>
